+++ test/test_nnls_merging_1Dparticles.jl
@@ -95,7 +95,8 @@
         append!(mim, compute_multi_index_moments(i))
     end
 
-    mnnls = NNLSMerge(mim, 30)
+    mnnls = NNLSMerge{1}(mim, 30)
+
     vref = sqrt(2 * k_B * 300.0 / species_data[1].mass)
     result = merge_nnls_based!(rng, mnnls, particles[1], pia, 1, 1; vref=vref, scaling=:vref)
     @test length(mnnls.rhs_vector) == length(mim) + 1  # we didn't construct 0-th order moment
@@ -116,13 +117,6 @@
     # test some internal computes used in the merging
     @test sum(abs.(mnnls.v0 .- v0)) < eps()
     @test sum(abs.(mnnls.w_total .- n0)) < 1.5e-14
-
-    @test abs(mnnls.minvx + v0[1] - 1.0) < eps()
-    @test abs(mnnls.maxvx + v0[1] - 50.0) < eps()
-    @test abs(mnnls.minvy + v0[2] - (5.0 - 2.0 * 50)) < eps()
-    @test abs(mnnls.maxvy + v0[2] - (5.0 - 2.0 * 1)) < eps()
-    @test abs(mnnls.minvz + v0[3] - (-100 + 0.25)) < eps()
-    @test abs(mnnls.maxvz + v0[3] - (-100 + 0.25 * 2500)) < eps()
 
     compute_props!(particles, pia, species_data, phys_props)
     # test that merging conserves mass / momentum / energy
@@ -150,7 +144,7 @@
         append!(mim, compute_multi_index_moments(i))
     end
 
-    mnnls2 = NNLSMerge(mim, 30)
+    mnnls2 = NNLSMerge{1}(mim, 30)
     particles2 = [create_particles2()]
     pia2 = ParticleIndexerArray(length(particles2[1]))
     vref = 1.0
@@ -170,7 +164,7 @@
 
 
     # check we always preserve the lowest-order moments
-    mnnls = NNLSMerge([], 30)
+    mnnls = NNLSMerge{1}([], 30)
     @test [0,0,0] in mnnls.mim
     @test [1,0,0] in mnnls.mim
     @test [0,1,0] in mnnls.mim
@@ -189,19 +183,12 @@
     particles = [create_particles3()]
     pia = ParticleIndexerArray(length(particles[1]))
 
-    centered_at_mean = true
-    v_multipliers = [0.25, 0.5, 1.0]
-    n_add = centered_at_mean ? 1 : 0
-    n_rand_pairs = 0
-    n_add += 8 * length(v_multipliers)
-    lhs_ncols = pia.indexer[1, 1].n_local + n_add + n_rand_pairs
+    lhs_ncols = pia.indexer[1, 1].n_local
     lhs_matrix = zeros(mnnls.n_moments_vel, lhs_ncols)
     vel_pos_matrix = zeros(6, lhs_ncols)
 
     Merzbild.compute_lhs_and_rhs!(mnnls, lhs_matrix, vel_pos_matrix,
                                   particles[1], pia, 1, 1)
-    # we didn't compute anything using the additional particles
-    @test sum(abs.(lhs_matrix[:, pia.indexer[1, 1].n_local+1:end])) == 0
 
     max_v_diff = 0.0
     max_x_diff = 0.0
@@ -211,8 +198,6 @@
                          abs.(vel_pos_matrix[3,i] - particles[1][i].v[3]),
                          max_v_diff)
         max_x_diff = max(abs.(vel_pos_matrix[4,i] - particles[1][i].x[1]),
-                         abs.(vel_pos_matrix[5,i] - particles[1][i].x[2]),
-                         abs.(vel_pos_matrix[6,i] - particles[1][i].x[3]),
                          max_x_diff)
     end
     @test max_v_diff <= 2*eps()
@@ -222,7 +207,7 @@
     #  [[0, 0, 0],
     # [1, 0, 0], [0, 1, 0], [0, 0, 1],
     # [2, 0, 0], [0, 2, 0], [0, 0, 2]]
-    mnnls3 = NNLSMerge(mim, 30)
+    mnnls3 = NNLSMerge{1}(mim, 30)
 
     lhs_ncols = 2
     # 7, 
@@ -241,9 +226,7 @@
     mnnls3.inv_vref = 1.0 / mnnls3.vref
     Merzbild.scale_lhs_rhs!(mnnls3, lhs_matrix, :vref, size(lhs_matrix, 2))
 
-    @test abs(mnnls3.scalevx - 2.0) < 2*eps()
-    @test abs(mnnls3.scalevy - 2.0) < 2*eps()
-    @test abs(mnnls3.scalevz - 2.0) < 2*eps()
+    @test maximum(abs.(mnnls3.scalev - [2.0, 2.0, 2.0])) < 2*eps()
     
     @test abs(rhs_vec_2[1] - mnnls3.rhs_vector[1]) < 2*eps()
     @test abs(rhs_vec_2[2] / mnnls3.vref - mnnls3.rhs_vector[2]) < 2*eps()
@@ -266,9 +249,7 @@
     mnnls3.rhs_vector = copy(rhs_vec_2)
     lhs_matrix = copy(lhs_matrix_2)
 
-    mnnls3.Ex = 2.0
-    mnnls3.Ey = 3.0
-    mnnls3.Ez = 4.0
+    mnnls3.Ev = SVector{3,Float64}(2.0, 3.0, 4.0)
 
     Merzbild.scale_lhs_rhs!(mnnls3, lhs_matrix, :variance, size(lhs_matrix, 2))
 
@@ -280,10 +261,7 @@
     @test abs(rhs_vec_2[6] / 9.0 - mnnls3.rhs_vector[6]) < 2*eps()
     @test abs(rhs_vec_2[7] / 16.0 - mnnls3.rhs_vector[7]) < 2*eps()
 
-    @test abs(mnnls3.scalevx - 2.0) < 2*eps()
-    @test abs(mnnls3.scalevy - 3.0) < 2*eps()
-    @test abs(mnnls3.scalevz - 4.0) < 2*eps()
-
+    @test maximum(abs.(mnnls3.scalev - [2.0, 3.0, 4.0])) < 2*eps()
     for i in 1:2
         @test maximum(abs.(lhs_matrix_2[1,i] - lhs_matrix[1,i])) < 2*eps()
         @test maximum(abs.(lhs_matrix_2[2,i] / 2.0 - lhs_matrix[2,i])) < 2*eps()
@@ -299,11 +277,13 @@
     particles = [create_particles2(mult=2.0,mult2=2.0)]
    
     pos_moments = [[0,0,1],[1,0,0],[0,1,0]]
+    # actual moments inside will be just [1,0,0]
 
-    mnnls_pos = NNLSMerge([], 30; multi_index_moments_pos=pos_moments)
-    @test mnnls_pos.pos_i_x == 2
-    @test mnnls_pos.pos_i_y == 3
-    @test mnnls_pos.pos_i_z == 1
+    mnnls_pos = NNLSMerge{1}([], 30; multi_index_moments_pos=pos_moments)
+    @test mnnls_pos.mim_pos == [[1,0,0]]
+    @test mnnls_pos.pos_i_x == 1
+    @test mnnls_pos.pos_i_y == -1
+    @test mnnls_pos.pos_i_z == -1
 
     x_mean = [0.0]
     v_mean = [0.0, 0.0, 0.0]
@@ -330,7 +310,7 @@
     pia = ParticleIndexerArray(length(particles[1]))
     result = merge_nnls_based!(rng, mnnls_pos, particles[1], pia, 1, 1)
     
-    @test length(mnnls_pos.rhs_vector) == 10
+    @test length(mnnls_pos.rhs_vector) == 8
     @test mnnls_pos.n_total_conserved == mnnls_pos.n_moments_vel + mnnls_pos.n_moments_pos
     
     @test result == 1
@@ -339,18 +319,18 @@
 
     n_new = pia.indexer[1,1].n_local[1]
 
-    @test maximum(abs.([mnnls_pos.Epx, mnnls_pos.Epy, mnnls_pos.Epz] - x_std)) < 4*eps()
-    @test maximum(abs.([mnnls_pos.scalex, mnnls_pos.scaley, mnnls_pos.scalez] - x_std)) < 4*eps()
+    @test maximum(abs.(mnnls_pos.Ex - x_std)) < 4*eps()
+    @test maximum(abs.(mnnls_pos.scalex - x_std)) < 4*eps()
 
     # we only conserve mean position
-    @test maximum(abs.(mnnls_pos.rhs_vector[8:10])) < 4*eps() 
+    @test maximum(abs.(mnnls_pos.rhs_vector[8:8])) < 4*eps() 
     
     new_w = 0.0
     new_v = [0.0, 0.0, 0.0]
     new_x = [0.0]
     s1 = pia.indexer[1,1].start1
     e1 = pia.indexer[1,1].end1
-    for i in 1:s1:e1
+    for i in s1:e1
         new_w += particles[1][i].w
         new_v += particles[1][i].v * particles[1][i].w
         new_x += particles[1][i].x * particles[1][i].w
@@ -359,7 +339,7 @@
     if pia.indexer[1,1].n_group2 > 0
         s2 = pia.indexer[1,1].start2
         e2 = pia.indexer[1,1].end2
-        for i in 1:s2:e2
+        for i in s2:e2
             new_w += particles[1][i].w
             new_v += particles[1][i].v * particles[1][i].w
             new_x += particles[1][i].x * particles[1][i].w
@@ -387,7 +367,7 @@
     new_x = [0.0]
     s1 = pia.indexer[1,1].start1
     e1 = pia.indexer[1,1].end1
-    for i in 1:s1:e1
+    for i in s1:e1
         new_w += particles[1][i].w
         # new_v += particles[1][i].v * particles[1][i].w
         # new_x += particles[1][i].x * particles[1][i].w
@@ -396,7 +376,7 @@
     if pia.indexer[1,1].n_group2 > 0
         s2 = pia.indexer[1,1].start2
         e2 = pia.indexer[1,1].end2
-        for i in 1:s2:e2
+        for i in s2:e2
             new_w += particles[1][i].w
             # new_v += particles[1][i].v * particles[1][i].w
             # new_x += particles[1][i].x * particles[1][i].w
@@ -413,23 +393,24 @@
     # test that computation of spatial moments does not affect the LHS and RHS entries corresponding to the velocity moments
     particles = [create_particles2(mult=2.0,mult2=2.0)]
     pia = ParticleIndexerArray(length(particles[1]))
-    lhs_ncols = 16 + 1 + 16  # 16 particles + 1 centered at 0 + 16 in octants
+    lhs_ncols = 16
    
     add_vel_moments = [[3,1,0],[0,2,1]]
     pos_moments = [[1,0,0],[0,1,0],[2,0,2]]
 
-    mnnls_pos = NNLSMerge(add_vel_moments, 30; multi_index_moments_pos=pos_moments)
-    mnnls_no_pos = NNLSMerge(add_vel_moments, 30)
+    # actual position moments will be [[1,0,0], [2,0,2]]
+    mnnls_pos = NNLSMerge{1}(add_vel_moments, 30; multi_index_moments_pos=pos_moments)
+    mnnls_no_pos = NNLSMerge{1}(add_vel_moments, 30)
     @test mnnls_pos.n_total_conserved == mnnls_pos.n_moments_vel + mnnls_pos.n_moments_pos
     @test mnnls_no_pos.n_total_conserved == mnnls_no_pos.n_moments_vel + mnnls_no_pos.n_moments_pos
 
     @test mnnls_pos.n_moments_vel == mnnls_no_pos.n_moments_vel
     @test mnnls_no_pos.n_moments_pos == 0
-    @test mnnls_pos.n_moments_pos == 3
+    @test mnnls_pos.n_moments_pos == 2
     @test mnnls_pos.pos_i_x == 1
-    @test mnnls_pos.pos_i_y == 2
+    @test mnnls_pos.pos_i_y == -1
     @test mnnls_pos.pos_i_z == -1
-    @test length(mnnls_pos.rhs_vector) == length(mnnls_no_pos.rhs_vector) + 3 == 12
+    @test length(mnnls_pos.rhs_vector) == length(mnnls_no_pos.rhs_vector) + 2 == 11
 
     lhs_matrix_pos = zeros(mnnls_pos.n_moments_vel + mnnls_pos.n_moments_pos, lhs_ncols)
     lhs_matrix_no_pos = zeros(mnnls_no_pos.n_moments_vel, lhs_ncols)
@@ -445,18 +426,6 @@
     # check computation of central spatial moments
     for i in 1:16
         @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+1,i] - (particles[1][i].x[1] - mnnls_pos.x0[1])) < 2*eps()
-        @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+2,i] - (particles[1][i].x[2] - mnnls_pos.x0[2])) < 2*eps()
-        @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+3,i] - (particles[1][i].x[1] - mnnls_pos.x0[1])^2 * (particles[1][i].x[3] - mnnls_pos.x0[3])^2) < 2*eps()
-    end
-
-    old_val = lhs_matrix_pos[mnnls_pos.n_moments_vel+2,3]
-    old_val2 = lhs_matrix_pos[mnnls_pos.n_moments_vel+3,3]
-
-    # additional fictitious particles
-    for i in 17:16+17
-        @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+1,i] - (0.0)) < 2*eps()
-        @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+2,i] - (0.0)) < 2*eps()
-        @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+3,i] - (0.0)^2 * (0.0)^2) < 2*eps()
     end
 
     # scaling of spatial moments also doesn't interfere with scaling of velocity moments
@@ -466,28 +435,11 @@
     @test maximum(abs.(mnnls_pos.rhs_vector[1:mnnls_pos.n_moments_vel] - mnnls_no_pos.rhs_vector)) < 2*eps()
     @test maximum(abs.(lhs_matrix_pos[1:mnnls_pos.n_moments_vel,:] - lhs_matrix_no_pos)) < 2*eps()
 
-    # test correct scaling of spatial moments
-    @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+2,3] * x_std[2] - old_val) < 2*eps()
-    @test abs(lhs_matrix_pos[mnnls_pos.n_moments_vel+3,3] * x_std[1]^2 * x_std[3]^2 - old_val2) < 2*eps()
-
     # finally check that we write 0.0 + mean_position to the position data in case 1st moment not specified
     result = merge_nnls_based!(rng, mnnls_pos, particles[1], pia, 1, 1; w_threshold=1e-12)
     @test result == 1
-    s1 = pia.indexer[1,1].start1
-    e1 = pia.indexer[1,1].end1
-    for i in 1:s1:e1
-        @test abs(particles[1][i].x[3] - mnnls_pos.x0[3]) <= eps()
-    end
 
-    if pia.indexer[1,1].n_group2 > 0
-        s2 = pia.indexer[1,1].start2
-        e2 = pia.indexer[1,1].end2
-        for i in 1:s2:e2
-            @test abs(particles[1][i].x[3] - mnnls_pos.x0[3]) <= eps()
-        end
-    end
-
-    mnnls_pos = NNLSMerge(add_vel_moments, 25; multi_index_moments_pos=pos_moments, matrix_ncol_nprealloc=10)
+    mnnls_pos = NNLSMerge{1}(add_vel_moments, 25; multi_index_moments_pos=pos_moments, matrix_ncol_nprealloc=10)
     @test length(mnnls_pos.lhs_matrices) == 11
     @test length(mnnls_pos.work) == 12
     @test mnnls_pos.lhs_matrix_ncols_start == 25
@@ -502,48 +454,43 @@
         @test size(mnnls_pos.work[i-25+1].idx) == (i,)
     end
 
+    # test merging with pre-allocated NNLS work arrays
+    particles = [create_particles2(mult=2.0,mult2=2.0)]
+    pia = ParticleIndexerArray(length(particles[1]))
     result = merge_nnls_based!(rng, mnnls_pos, particles[1], pia, 1, 1; w_threshold=1e-12)
     @test result == 1
-    s1 = pia.indexer[1,1].start1
-    e1 = pia.indexer[1,1].end1
-    for i in 1:s1:e1
-        @test abs(particles[1][i].x[3] - mnnls_pos.x0[3]) <= eps()
-    end
-
-    if pia.indexer[1,1].n_group2 > 0
-        s2 = pia.indexer[1,1].start2
-        e2 = pia.indexer[1,1].end2
-        for i in 1:s2:e2
-            @test abs(particles[1][i].x[3] - mnnls_pos.x0[3]) <= eps()
-        end
-    end
 
     particles = [create_particles2(mult=2.0,mult2=2.0)]
     pia = ParticleIndexerArray(length(particles[1]))
-    result = merge_nnls_based!(rng, mnnls_pos, particles[1], pia, 1, 1; centered_at_mean=false, n_rand_pairs=3, w_threshold=1e-12)
+    result = merge_nnls_based!(rng, mnnls_pos, particles[1], pia, 1, 1; w_threshold=1e-12)
     @test result == 1
+    @test pia.index_last[1] == pia.n_total[1]
+
     wtot = 0.0
     s1 = pia.indexer[1,1].start1
     e1 = pia.indexer[1,1].end1
-    for i in 1:s1:e1
+    for i in s1:e1
         wtot += particles[1][i].w
     end
 
     if pia.indexer[1,1].n_group2 > 0
         s2 = pia.indexer[1,1].start2
         e2 = pia.indexer[1,1].end2
-        for i in 1:s2:e2
+        for i in s2:e2
             wtot += particles[1][i].w
         end
     end
 
-    @test abs(new_w - tw) < 4.5e-15
+    @test abs(wtot - tw) < 4.5e-15
 
     # finally test the edge case of init_np = total_np and no pre-allocated matrices, v_multipliers=[]
     # reset particles
     particles = [create_particles2(mult=2.0,mult2=2.0)]
     pia = ParticleIndexerArray(length(particles[1]))
 
-    mnnls = NNLSMerge(add_vel_moments, length(particles[1]); multi_index_moments_pos=[])
-    merge_nnls_based!(rng, mnnls, particles[1], pia, 1, 1; centered_at_mean=false, n_rand_pairs=0, v_multipliers=[], w_threshold=1e-12)
+    mnnls = NNLSMerge{1}(add_vel_moments, length(particles[1]); multi_index_moments_pos=[])
+    result = merge_nnls_based!(rng, mnnls, particles[1], pia, 1, 1; w_threshold=1e-12)
+
+    @test result == 1
+    @test pia.n_total[1] < 16
 end