+++ test/test_nnls_ratepreserving_merging.jl
@@ -61,7 +61,6 @@
                                                           Dict("Ar" => ScatteringIsotropic),
                                                           Dict("Ar" => ElectronEnergySplitEqual))
 
-
     # we do a trick here
     # we don't have actual neutral particles
     # so we can't set `species` to 3 (electrons), because indexing gets messed up
@@ -134,8 +133,8 @@
                                                interaction_data, n_e_interactions, computed_cs,
                                                particles, pia, 1, 1, 1,
                                                cs_ref, cs_ref; scaling=:variance,
-                                               vref=vref, n_rand_pairs=0, max_err=1e-11,
-                                               centered_at_mean=false, v_multipliers=[], iteration_mult=2,
+                                               vref=vref, max_err=1e-11,
+                                               iteration_mult=2,
                                                extend=CSExtendConstant)
 
     @test result == 1
@@ -147,8 +146,8 @@
     k_rate_ionization = rate_ionization / w0
 
     # test RHS rate coefficients
-    scaled_k_elastic = k_rate_elastic / (cs_ref * sqrt(nnls_rp.Ex^2 + nnls_rp.Ey^2 + nnls_rp.Ez^2))
-    scaled_k_ion = k_rate_ionization / (cs_ref * sqrt(nnls_rp.Ex^2 + nnls_rp.Ey^2 + nnls_rp.Ez^2))
+    scaled_k_elastic = k_rate_elastic / (cs_ref * sqrt(nnls_rp.Ev[1]^2 + nnls_rp.Ev[2]^2 + nnls_rp.Ev[3]^2))
+    scaled_k_ion = k_rate_ionization / (cs_ref * sqrt(nnls_rp.Ev[1]^2 + nnls_rp.Ev[2]^2 + nnls_rp.Ev[3]^2))
 
     @test abs(nnls_rp.rhs_vector[8] - scaled_k_elastic)/scaled_k_elastic < 4*eps()
     @test abs(nnls_rp.rhs_vector[9] - scaled_k_ion)/scaled_k_ion < 4*eps()
@@ -223,8 +222,8 @@
                                                interaction_data, n_e_interactions, computed_cs,
                                                particles, pia, 1, 1, 1,
                                                cs_ref, cs_ref; scaling=:variance,
-                                               vref=vref, n_rand_pairs=0, max_err=1e-11,
-                                               centered_at_mean=false, v_multipliers=[], iteration_mult=2,
+                                               vref=vref, max_err=1e-11,
+                                               iteration_mult=2,
                                                extend=CSExtendConstant)
 
     @test result == 1
@@ -237,7 +236,7 @@
 
     # test RHS rate coefficients
 
-    scaled_k_elastic = k_rate_elastic / (cs_ref * sqrt(nnls_rp.Ex^2 + nnls_rp.Ey^2 + nnls_rp.Ez^2))
+    scaled_k_elastic = k_rate_elastic / (cs_ref * sqrt(nnls_rp.Ev[1]^2 + nnls_rp.Ev[2]^2 + nnls_rp.Ev[3]^2))
 
     @test abs(nnls_rp.rhs_vector[8] - scaled_k_elastic)/scaled_k_elastic < 4*eps()
 
@@ -277,7 +276,7 @@
     @test abs(rate_elastic_new - rate_elastic)/rate_elastic < 1e-15
     @test abs(rate_ionization_new - rate_ionization) < 1e-15
 
-    # finally test the edge case of init_np = total_np and no pre-allocated matrices, v_multipliers=[]
+    # finally test the edge case of init_np = total_np
     # reset particles
     particles, pia = create_particles(ndens)
 
@@ -287,8 +286,8 @@
                                                interaction_data, n_e_interactions, computed_cs,
                                                particles, pia, 1, 1, 1,
                                                cs_ref, cs_ref; scaling=:variance,
-                                               vref=vref, n_rand_pairs=0, max_err=1e-11,
-                                               centered_at_mean=false, v_multipliers=[], iteration_mult=2,
+                                               vref=vref, max_err=1e-11,
+                                               iteration_mult=2,
                                                extend=CSExtendConstant)
 
     @test result == 1